SQL-Agent/requirements.txt
langchain
langchain_community 
langchain_openai
streamlit 
python-dotenv